+++ tests/test_checker/test_noqa.py
@@ -40,6 +40,7 @@ def test_noqa_fixture_disabled(absolute_path, all_violations):
         'Z114': 1,
         'Z115': 1,
         'Z116': 1,
+        'Z117': 1,
 
         'Z200': 0,
         'Z201': 0,
@@ -100,6 +101,8 @@ def test_noqa_fixture_disabled(absolute_path, all_violations):
         'Z440': 1,
         'Z441': 1,
         'Z442': 1,
+        'Z443': 1,
+        'Z444': 1,
     }
 
     process = subprocess.Popen(