+++ tests/test_visitors/test_ast/test_naming/test_naming_rules/test_consecutive_underscore.py
@@ -1,0 +1,32 @@
+# -*- coding: utf-8 -*-
+
+import pytest
+
+from wemake_python_styleguide.violations.naming import (
+    ConsecutiveUnderscoresInNameViolation,
+)
+from wemake_python_styleguide.visitors.ast.naming import WrongNameVisitor
+
+
+@pytest.mark.parametrize('underscored_name', [
+    'with__underscore',
+    'mutliple__under__score',
+    'triple___underscore',
+])
+def test_underscored_variable_name(
+    assert_errors,
+    assert_error_text,
+    parse_ast_tree,
+    naming_template,
+    default_options,
+    mode,
+    underscored_name,
+):
+    """Ensures that underscored names are not allowed."""
+    tree = parse_ast_tree(mode(naming_template.format(underscored_name)))
+
+    visitor = WrongNameVisitor(default_options, tree=tree)
+    visitor.run()
+
+    assert_errors(visitor, [ConsecutiveUnderscoresInNameViolation])
+    assert_error_text(visitor, underscored_name)