+++ tests/test_visitors/test_ast/test_naming/test_naming_rules/test_restricted_argument.py
@@ -1,0 +1,35 @@
+# -*- coding: utf-8 -*-
+
+import pytest
+
+from wemake_python_styleguide.constants import SPECIAL_ARGUMENT_NAMES_WHITELIST
+from wemake_python_styleguide.violations.naming import (
+    ReservedArgumentNameViolation,
+)
+from wemake_python_styleguide.visitors.ast.naming import WrongNameVisitor
+
+
+@pytest.mark.parametrize('argument', SPECIAL_ARGUMENT_NAMES_WHITELIST)
+def test_restricted_argument_name(
+    assert_errors,
+    assert_error_text,
+    parse_ast_tree,
+    naming_template,
+    default_options,
+    mode,
+    argument,
+):
+    """Ensures that special names for arguments are restricted."""
+    if argument in naming_template:
+        # We are facing a strange behavior here,
+        # since `ast.parse()` works well with two same arguments.
+        # But it should not!
+        return
+
+    tree = parse_ast_tree(mode(naming_template.format(argument)))
+
+    visitor = WrongNameVisitor(default_options, tree=tree)
+    visitor.run()
+
+    assert_errors(visitor, [ReservedArgumentNameViolation])
+    assert_error_text(visitor, argument)