+++ wemake_python_styleguide/violations/best_practices.py
@@ -42,6 +42,8 @@ Summary
    ProtectedModuleViolation
    ProtectedAttributeViolation
    LambdaInsideLoopViolation
+   UnusedArgumentViolation
+   UnusedArgumentIsUsedViolation
 
 Comments
 --------
@@ -80,6 +82,8 @@ Design
 .. autoclass:: ProtectedModuleViolation
 .. autoclass:: ProtectedAttributeViolation
 .. autoclass:: LambdaInsideLoopViolation
+.. autoclass:: UnusedArgumentViolation
+.. autoclass:: UnusedArgumentIsUsedViolation
 
 """
 
@@ -882,3 +886,83 @@ class LambdaInsideLoopViolation(ASTViolation):
     should_use_text = False
     error_template = "Found `lambda` in loop's body"
     code = 442
+
+
+@final
+class UnusedArgumentViolation(ASTViolation):
+    """
+    Forbids to have unused arguments in functions and methods.
+
+    Reasoning:
+        Having unused arguments is confusing. Why do we have them?
+        Is it some kind of API we must stick to? Or is it just an error?
+
+    Solution:
+        There are two possible solutions.
+        Firstly, try to remove this argument if you can.
+        Secondly, you can add a leading underscore to the argument's name.
+
+    Example::
+
+        # Correct:
+        def function(first):
+            return first + 10
+
+        def function(first, _second):
+            return first + 10
+
+        # Wrong:
+        def function(first, second):
+            return first + 10
+
+    This rule checks: functions, methods, and ``lambda`` functions.
+    We also have an exception for special arguments like:
+    ``self``, ``cls``, and ``mcs``.
+    These arguments can be unused inside defined methods.
+
+    .. versionadded:: 0.5.0
+
+    """
+
+    error_template = 'Found unused argument: {0}'
+    code = 443
+
+
+@final
+class UnusedArgumentIsUsedViolation(ASTViolation):
+    """
+    Forbids to have use arguments that are marked as unused.
+
+    Reasoning:
+        Sometimes your start to use new logic in your functions,
+        and you start to use arguments that once were marked as unused.
+        But, you have not renamed them for some reason.
+        And now you have a lot of confusion: argument is marked as unused,
+        but you are using it. Why? What's going on?
+        This rule is heavy related with :class:`UnusedArgumentViolation`.
+
+    Solution:
+        Rename your argument to be a regular argument,
+        in other words: remove the leading underscore.
+
+    Example::
+
+        # Correct:
+        def function(first):
+            return first + 10
+
+        def function(first, _second):
+            return first + 10
+
+        # Wrong:
+        def function(first, _second):
+            return first + _second
+
+    This rule checks: functions, methods, and ``lambda`` functions.
+
+    .. versionadded:: 0.5.0
+
+    """
+
+    error_template = 'Found using argument marked as unused: {0}'
+    code = 444