+++ wemake_python_styleguide/visitors/ast/imports.py
@@ -21,44 +21,44 @@ from wemake_python_styleguide.violations.consistency import (
 from wemake_python_styleguide.violations.naming import SameAliasImportViolation
 from wemake_python_styleguide.visitors.base import BaseNodeVisitor
 
-ErrorCallback = Callable[[BaseViolation], None]
+ErrorCallback = Callable[[BaseViolation], None]  # TODO: alias and move
 
 
 @final
-class _ImportsChecker(object):
+class _ImportsValidator(object):
     """Utility class to separate logic from the visitor."""
 
     def __init__(self, error_callback: ErrorCallback) -> None:
-        self.error_callback = error_callback
+        self._error_callback = error_callback
 
     def check_nested_import(self, node: AnyImport) -> None:
         parent = getattr(node, 'parent', None)
         if parent is not None and not isinstance(parent, ast.Module):
-            self.error_callback(NestedImportViolation(node))
+            self._error_callback(NestedImportViolation(node))
 
     def check_local_import(self, node: ast.ImportFrom) -> None:
         if node.level != 0:
-            self.error_callback(LocalFolderImportViolation(node))
+            self._error_callback(LocalFolderImportViolation(node))
 
     def check_future_import(self, node: ast.ImportFrom) -> None:
         if node.module == '__future__':
             for alias in node.names:
                 if alias.name not in FUTURE_IMPORTS_WHITELIST:
-                    self.error_callback(
+                    self._error_callback(
                         FutureImportViolation(node, text=alias.name),
                     )
 
     def check_dotted_raw_import(self, node: ast.Import) -> None:
         for alias in node.names:
             if '.' in alias.name:
-                self.error_callback(
+                self._error_callback(
                     DottedRawImportViolation(node, text=alias.name),
                 )
 
     def check_alias(self, node: AnyImport) -> None:
         for alias in node.names:
             if alias.asname == alias.name:
-                self.error_callback(
+                self._error_callback(
                     SameAliasImportViolation(node, text=alias.name),
                 )
 
@@ -66,7 +66,7 @@ class _ImportsChecker(object):
         import_names = [alias.name for alias in node.names]
         for name in chain(imports.get_import_parts(node), import_names):
             if access.is_protected(name):
-                self.error_callback(ProtectedModuleViolation(node))
+                self._error_callback(ProtectedModuleViolation(node))
 
 
 @final
@@ -76,7 +76,7 @@ class WrongImportVisitor(BaseNodeVisitor):
     def __init__(self, *args, **kwargs) -> None:
         """Creates a checker for tracked violations."""
         super().__init__(*args, **kwargs)
-        self._checker = _ImportsChecker(self.add_violation)
+        self._validator = _ImportsValidator(self.add_violation)
 
     def visit_Import(self, node: ast.Import) -> None:
         """
@@ -88,10 +88,10 @@ class WrongImportVisitor(BaseNodeVisitor):
             NestedImportViolation
 
         """
-        self._checker.check_nested_import(node)
-        self._checker.check_dotted_raw_import(node)
-        self._checker.check_alias(node)
-        self._checker.check_protected_import(node)
+        self._validator.check_nested_import(node)
+        self._validator.check_dotted_raw_import(node)
+        self._validator.check_alias(node)
+        self._validator.check_protected_import(node)
         self.generic_visit(node)
 
     def visit_ImportFrom(self, node: ast.ImportFrom) -> None:
@@ -105,9 +105,9 @@ class WrongImportVisitor(BaseNodeVisitor):
             FutureImportViolation
 
         """
-        self._checker.check_local_import(node)
-        self._checker.check_nested_import(node)
-        self._checker.check_future_import(node)
-        self._checker.check_alias(node)
-        self._checker.check_protected_import(node)
+        self._validator.check_local_import(node)
+        self._validator.check_nested_import(node)
+        self._validator.check_future_import(node)
+        self._validator.check_alias(node)
+        self._validator.check_protected_import(node)
         self.generic_visit(node)